Add support to enable singleton bucket in histogram
If gpdb and orca is build with debugging enabled, there is an assert in
orca to check if the upper and lower bound are both closed.
`GPOS_ASSERT_IMP(FSingleton(), fLowerClosed && fUpperClosed);`

The histogram that is stored in pg_statistics might lead to have a
singleton buckets as follows: `10, 20, 20, 30, 40` which will lead to have
buckets in this format: `[0,10), [10, 20), [20,20), [20,30), [30,40]`

This will cause assert to fail since [20,20) is a singleton bucket but
its upper bound is open.

With this fix, the generated buckets will look like below:
`[0,10], [10,20), [20,20], (20,30), [30,40]`

Signed-off-by: Shreedhar Hardikar <shardikar@pivotal.io>