Specify the instance size required in the test
The instance size specified in the ccp_options anchor do not have to
match in the ccp_destroy anchor, the only requirement is that the field
must exist. With this revelation, there does not need to be test
specific anchors defined in ccp_options.

Instead the pattern is to not use a ccp_options anchor, but instead to
explicitly set the desired instance type for a given test.

Signed-off-by: Jingyi Mei <jmei@pivotal.io>