Suggested changes from toolsmiths
The source code for gpdb is being copied to the master node, so there is no need to run scripts
in the container that can be directly run on the master node.  This makes scripts in the GPDB source
tree not depend on the connection that CI system has to the cluster, just that your test is running on
a cluster.

Signed-off-by: Divya Bhargov <dbhargov@pivotal.io>