Install libevent in Travis CI builds
Since Travis CI was upgraded to use Ubuntu Trusty as the base
Linux [0], libevent-dev is no longer available out of the box.
Install it manually with the Apt addon, since we don't want to
use sudo due to longer instance bootup time. While at it,
remove the macOS section since we never actually supported
Travis for macOS builds (it was a leftover from an attempt).

[0] https://blog.travis-ci.com/2017-07-11-trusty-as-default-linux-is-coming