Testing for full page image after hint bit change.
With `gp_disable_tuple_hints` off, buffer will be marked dirty for hint
bit changes. The XLOG should capture the full page image if the hint bit
is the first change after checkpoint.

During recovery, the full page image should be replayed to override the
page, even if the page is corrupted on disk.

This test verified the table with corrupted page full fixed after
recovery if the `gp_disable_tuple_hints` off.

Signed-off-by: Xin Zhang <xzhang@pivotal.io>