gpinitsystem: Fix gpinitsystem -I with mixed case names
We are hitting an issue where gpinitsystem sends a gpstop command that
has the data directory in all lower case. This causes an issue in
Linux, but is fine with macOS; linux platforms are case sensitive vs
macOS is not.

Remove the lower casing part of the code as it doesn't seem necessary.
gpinitsystem -c command eventually overwrites the SEG_PREFIX variable in a
different function to use the mixed case, while the gpinitsystem -I does not.

Fix: keep the mixed case segment names