Protect CCP users from a breaking change
In order to protect users of the concourse cluster provisioner service from
some pending breaking changes, we are pinning the master pipeline to the latest
tag for the concourse cluster provisioner repo.  We will in the future facilitate
upgrades on the master pipeline to take advantage of the next release.

Signed-off-by: Jim Doty <jdoty@pivotal.io>