Replace low-level mock test with pg_regress regression test.
One less test program makes the tests to go a tiny bit faster. This also
tests more directly the problem with the GUCs that the low-level function
was added for, rather than the low-level function.

Arguably we don't need a test for any of this at all anymore, as this fix
for the ALTER USER RESET ALL bug was backported from PostgreSQL a long time
ago, but at this point, we have caught up with that PostgreSQL code and
the GPDB code is identical to the upstream. But OTOH, it's easy and quick to
test as a pg_regress test, so might as well.