Create minimal segment WAL replication pipeline
As we develop segment WAL replication, we need a green to green CI
pipeline that we can start small and add relative jobs from
gpdb_master as we add more functionality. This commit establishes that
pipeline and omits ICW tests that are currently not supported by
segment WAL replication.

Authors: Abhijit Subramanya and Jimmy Yih