gprecoverseg: validate checksum setting
As part of the validation phase of gprecoverseg, before proceeding,
validate that the setting for GUC data_checksums is the same between
master and segments. The validation is done by comparing pg_control file content.
Fail fast if the settings are not the same.

If no segments are able to report their settings, then gprecoverseg
fails. (This failure to report would be unexpected since there is already a check for at least
one segment alive to progress to the validation phase.)

Signed-off-by: Marbin Tan <mtan@pivotal.io>
Signed-off-by: Nadeem Ghani <nghani@pivotal.io>