Match upstream code better in StoreAttrDefault function.
Upstream commit cd902b331dc, from 8.4, changed StoreAttrDefault function's
adbin argument from a string to a Node *. We had backported that, but the
code to deal with it inside the function was not 100% identical to the
upstream version. There was nothing wrong with the way it was written in
GPDB per se, but let's make it identical to the upstream version anyway,
to avoid merge conflicts.