Replace uses of CopyRelPath with upstream relpath() function.
CopyRelPath() is essentially a copy of relpath(), but it returns the result
in a caller-supplied buffer instead of pallocing. That's redundant, and the
calls are not in performance-critical path, so let's just use relpath() and
get rid of CopyRelPath(). This reduces our diff footprint vs upstream. If
there were upstream changes to relpath(), we would surely forget to make
the same changes to CopyRelPath().