gpconfig: fix unit test failure
We were passing an empty list and appending it instead of extending it.
using an empty list as a default value for a function parameter is an
anti-pattern.
Python behaves unintuitively, where it creates a single instance of the
variable and uses it again.
Feels like how a static variable would behave in c.

Signed-off-by: Nadeem Ghani <nghani@pivotal.io>