Map \dx+ command to \dx, until PostgreSQL 9.1.
The \dx+ command requires pg_describe_object() function, which we don't
have in GPDB yet. Until then, treat \dx+ the same as \dx.

Note that this code is written such that \dx+ can still be used, when
connected to a PostgreSQL 9.1 or later server. Not that we particularly care
about that case, but might as well.

Per github issue #2574.