Fix extract_actual_clauses() calls.
Remove premature optimization in create_indexscan_plan, where we stripped
RestrictInfo decorations while we iterated through the list. That was
premature, because while it works, order_qual_clauses() can make use of the
information in the RestrictInfo, if it's still present. This reverts that
piece of code to the way it is in the upstream. I'm not sure if this makes
any observable difference, but let's be tidy in any case.

Also, reorder the extract_actual_clauses and order_qual_clauses calls
in create_ctescan_plan, so that order_qual_clauses is called first. For
the same reason.