Make resource group test cases pass when number of segments is not three
In some resource group test cases, the result will depend on the
number of segments, and current test cases are based on three
segments. This commit is to fix this issue.

In addition, in isolation2, gpdiff needs at least 3 dashes to identify 
the result of 'select' statement. Change the elements to be above 
100 in test case 'resgroup_memory_materialize_spill' to meet this 
requirement.