Use 64-bit integers for row counts in window functions.
Consistently use 64-bit integers for the variables and fields that hold
row counts, in the window functions code. These are used for the ROWS
PRECEDING / FOLLOWING clauses. I don't expect anyone to deal with more than
2 billion rows in a single query's window function, but better safe than
sorry. Also, upstream uses int64 for these, and bigint is also used for
OFFSET/LIMIT, so let's be consistent.