Auxiliary and some backend processes should be set to GP_ROLE_UTILITY
Currently, auxiliary and backend processes are defaulted to
GP_ROLE_DISPATCH.  This was kind of harmless until we started enabling
segment WAL replication.  The WAL sender would be initialized as a
backend process that would have a gp_session_id set and mppIsWriter be
True.  This would intermittently cause issues with cursors.  The writer
would open a temporary file with its session id and process id but the
reader would construct the temporary filename incorrectly with the
right session id but the WAL sender's process id.  The reader would
then fail to open the correct temporary file.

To fix this, we set Gp_role to GP_ROLE_UTILITY for all auxiliary
processes and make exceptions for certain backend processes
(Autovacuum and WAL sender).