Remove obsolete FIXME comments.
When the new OID-dispatching mechanism was introduced (commit f9016da2a6),
a lot of FIXME comments were left in calls to CdbDispatchUtilityStatement.
These were places where I wasn't sure if the command that was executed
might create new objects, with new OIDs assigned to them, that would need
to be dispatched to the QE. I've now skimmed through the call sites, and
checked that they are all for ALTER or DROP commands that should not create
new OIDs, so remove the FIXME comments.