Stop FTS probes for mirrorless cluster
An FTS probe to the primaries in a mirrorless cluster will never
result in the update of gp_segment_configuration.  If a primary goes
down, we must keep the primary marked as up so that gpstart can start
the primary back up.  All transactions will abort and nothing should
work except for read-only queries to master-only catalog tables.

Stopping the FTS probes for mirrorless cluster introduced an infinite
loop in FtsNotifyProber in which the dispatcher waits in an infinite
loop for fts_statusVersion to change.  To break the infinite loop, we
acknowledge the forced probe request as a no-op and update
fts_statusVersion to break the loop for the dispatcher.  The
dispatcher should then act same as before this commit.

We also add #define for character value of GpFaultStrategy.

Authors: Xin Zhang, Ashwin Agrawal, and Jimmy Yih