gpstart: Fix standby master pg_controldata heap checksum validation
Previously, we were only checking localhost for standby heap checksum in
gpstart. This is not the case as the standby is usually on a remote host.
gpstart raises an error due to not finding the pg_control file of standby locally.

Fixed by always doing a remote host call for the standby master.

Signed-off-by: Marbin Tan <mtan@pivotal.io>