Fix broken heap_form_tuple function.
I'm not sure if we broke this during the 8.3 merge, or if it was broken
even before that, but our version of heap_form_tuple was confused over
what the isnull argument is. The difference between heap_form_tuple()
and heap_formtuple() is the values used in the "isnull" array, but
our heap_form_tuple passed a 'char' array to the internal heap_fill_tuple()
function, which assumes a 'bool' arry.

That caused an assertion failure in hstore regression tests (and almost
certainly a crash or at least wrong results without assertions, though I
didn't check.)

To fix, take the version of heap_form_tuple() from PostgreSQL 8.4, which
converts the 'char' array to a 'bool' array, and then calls the new
heap_formtuple() function.