Fix Travis with --disable-pxf configure flag
Fix Travis by adding --disable-pxf to configure command. We could
always just add json-c-devel package and see if it works... but we
need to determine if we want Travis to compile pxf or not first. We
currently have Travis set up to not compile gpAux/extensions so this
PR just adds to the current setup.