Ensure error stack before using as array subscript
-1 is a valid value for errordata_stack_depth until errstart() has
been called which inits the stack. In correct usage it's thus fine
to use it as an array subscript without inspecting it, since demote
and dismiss are useless outside an error context. Incorrect usage
would however cause a memory error so guard against that since it's
the right thing to do.