Refactor TransformExttabAuthClause function slightly, for readability.
Seems better for the function to palloc the struct it returns, rather than
have it fill in a struct provided by the caller.

The thing that really caught my eye was this:

```
genericPair *genpair = (genericPair *) palloc0 (4 * sizeof(char *));
```

That wrorks, because the genericPair struct consists of four "char *"
fields. But it seems fragile, should use "sizeof(genericPair)". This commit
fixes that, among the other kibitzing.