Refactor ParseFuncOrColumn and func_get_detail to match upstream better.
Only ParseFuncOrColumn needed the additional "isstrict" and "isordered"
return values from func_get_detail. And even ParseFuncOrColumn only needed
that information under certain circumstances, for error checks. Move the
code to fetch that information out of func_get_detail, to separate
little helper functions.

In principle, fetching the "isstrict" and "isordered" flags separately is
more expensive, because it now requires two syscache lookups rather than
one. In practice, however, this is a win, because you only need to fetch
the "isstrict" flag if a FILTER clause was given, and you only need the
"isordered" flag if an OVER clause was given. Both of those clauses are
rare, so in the common case this actually saves one syscache lookup. (None
of this matters much in practice, though, because syscache lookups are
pretty cheap anyway.)