Don't initialize random seed when creating a temporary file.
That seems like a very random place to do it (sorry for the pun). The
random seed is initialized at backend startup anyway, that ought to be
good enough, so just remove the spurious initialization from bfz.c.

In the passing, improve the debug-message to mention which compression
algorithm was used.