Refactor code that selects a common type for columns in a UNION query.
The big difference is that each leaf query is now transformed in one go,
like it's done in the upstream, instead of transforming the target list
and FROM list first. That partial transformation was causing trouble for
another refactoring that I'm working on, which ill change the way window
functions are handled in parse analysis.

This two-pass code is GPDB-specific, PostgreSQL uses a simpler algorithm
that works bottom-up, one setop node at a time, to select the column types.