Decide whether a window edge is "delayed" later, in the executor.
In principle, it makes sense to determine at plan-time, whether the
expression needs to be re-evaluated for every row. In practice, it seems
simpler to decide that in the executor, when initializing the Window node.
This allows removing a bunch of code from the planner, and from the ORCA
translator, including the hack to force the expression to be delayed if it
was a SubLink.

The planner always set the delayed flag, unless the expression was a Const.
We can easily and quickly check for that in the executor too. I'm not sure
how ORCA decided whether to delay or not, but in some quick testing I
cannot come up with a case where it would decide differently.