Refactor EDGE_* macros in nodeWindow.c.
Refactor the EDGE_IS_* macros so that they take the current
WindowStatePerLevel as argument, rather than a WindowFrameEdge. This will
make the transition to the upstream window function implementation a bit
easier, because the upstream implementation doesn't have a WindowFrameEdge
struct.

Encapsulate access to level_state->is_rows in a macro. This is also in
preparation for the upstream implementation, which will contain that
information as a flag in the frameOptions bitmask.

Many functions used an 'is_lead' argument, and one used 'is_trail', to
keep track whether we're dealing with the leading or trailing edge.
Introduce a little enum with EDGE_TRAIL and EDGE_LEAD values, to avoid
having to remember what 'true' or 'false' means in which context.