Remove pessimization of trivial SubqueryScans over Window nodes.
This hack, to refrain from removing a trivial SubqueryScan if the subnode
was a Window node, was added back in 2009 along with the test case. I'm
not sure what the problem was, but this must've been just a quick band-aid
over whatever the real problem was. It doesn't seem to be needed anymore,
as everything works without it.

Expand the comment in the test case a little bit to explain what the point
of the query is. The original commit message said just "Fix MPP-4840", so
this is all the information I could find about it, unfortunately.