Optimize away DISTINCT and ORDER BY in a NOT IN (...) subselect.
It occurred to me while looking at PR #1460 that when there's a DISTINCT or
an ORDER BY in a subselect NOT IN (...) subselect won't make any difference
to the overall result, so we can strip it off and safe the effort. In its
current form, PR #1460 would pessimize that case slightly more, by forcing
the subselect's resul to be gathered to a singled node for deduplication or
final ordering, while before, we would do only a local ordering /
deduplication on each segment. But it is a waste of effort to do that
even within each segment, and this PR gets rid of that.