Refactor the removal of DISTINCT and ORDER BY, to reuse the existing code.
In commit d16710cada, I added an optimization to NOT IN subqueries, to
remove any DISTINCT and ORDER BY. But on second thoughts, we should use the
existing functions to do that. Also, the add_notin_subquery_rte() was not
a good place to do it. It's a surprising side-effect for that function.
Move the code to convert_IN_to_antijoin(), which is in line with the
similar calls in convert_EXPR_to_join() and convert_IN_to_join().