Map GPOS severity level to GPDB Severity Levels
GPOS raises exception with different severity level, but
they were being logged to GPDB logs at LOG severity level.

This is the initial commit which introduces the functionality. If an
exception is created with debug log level, these messages will be logged
to GPBD with debug1 level rest will be logged as LOG level.

Signed-off-by: Jemish Patel <jpatel@pivotal.io>