Fix remaining equal* functions to not compare 'location' field.
The 'location' field is just to give better error messages. It should not
be considered when testing whether two nodes are equal. (Note that the
COMPARE_LOCATION_FIELD() macro that we now consistently use on the
'location' field is a no-op.)

I noticed this while working on a patch that would compare two ColumnRefs
to see if they are equal, and could be collapsed to one.